app/azure_functions/finance_csv_ingestor/requirements.txt
azure-functions
azure-identity
azure-storage-blob
psycopg2-binary
pandas
requests
python-yahoofinance

